04_Connecting_LLM_Clients/langchain_mcp_adapters_project/pyproject.toml
[project]
name = "langchain-mcp-adapters-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.19",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
